HIVE-21164

+++ ql/src/java/org/apache/hadoop/hive/ql/exec/FileSinkOperator.java
@@ -161,6 +161,8 @@ public class FileSinkOperator extends TerminalOperator<FileSinkDesc> implements
     private final String subdirForTxn;
     private Path taskOutputTempPathRoot;
     Path[] outPaths;
+    // The bucket files we successfully wrote to in this writer
+    Path[] outPathsCommitted;
     Path[] finalPaths;
     RecordWriter[] outWriters;
     RecordUpdater[] updaters;
@@ -168,19 +170,29 @@ public class FileSinkOperator extends TerminalOperator<FileSinkDesc> implements
     int acidLastBucket = -1;
     int acidFileOffset = -1;
     private boolean isMmTable;
+    private boolean isDirectInsert;
     String dpDirForCounters;
 
-    public FSPaths(Path specPath, boolean isMmTable) {
-      this.isMmTable = isMmTable;
-      if (!isMmTable) {
+    public FSPaths(Path specPath, boolean isMmTable, boolean isDirectInsert) {
+      this.isMmTable = isMmTable; 
+      this.isDirectInsert = isDirectInsert;
+      if (!isMmTable && !isDirectInsert) {
         tmpPathRoot = Utilities.toTempPath(specPath);
         taskOutputTempPathRoot = Utilities.toTaskTempPath(specPath);
         subdirForTxn = null;
       } else {
         tmpPathRoot = specPath;
         taskOutputTempPathRoot = null; // Should not be used.
-        subdirForTxn = AcidUtils.baseOrDeltaSubdir(conf.getInsertOverwrite(),
-            conf.getTableWriteId(), conf.getTableWriteId(),  conf.getStatementId());
+        if (isMmTable) {
+          subdirForTxn = AcidUtils.baseOrDeltaSubdir(conf.getInsertOverwrite(),
+              conf.getTableWriteId(), conf.getTableWriteId(),  conf.getStatementId());
+        } else {
+          /**
+           * For direct write to final path during ACID insert, we create the delta directories
+           * later when we create the RecordUpdater using AcidOutputFormat.Options
+           */
+          subdirForTxn = null;
+        }
       }
       if (Utilities.FILE_OP_LOGGER.isTraceEnabled()) {
         Utilities.FILE_OP_LOGGER.trace("new FSPaths for " + numFiles
@@ -188,6 +200,7 @@ public class FileSinkOperator extends TerminalOperator<FileSinkDesc> implements
       }
 
       outPaths = new Path[numFiles];
+      outPathsCommitted = new Path[numFiles];
       finalPaths = new Path[numFiles];
       outWriters = new RecordWriter[numFiles];
       updaters = new RecordUpdater[numFiles];
@@ -211,7 +224,12 @@ public class FileSinkOperator extends TerminalOperator<FileSinkDesc> implements
       try {
         for (int i = 0; i < updaters.length; i++) {
           if (updaters[i] != null) {
-            updaters[i].close(abort);
+            SerDeStats stats = updaters[i].getStats();
+            // Ignore 0 row files
+            if (isDirectInsert && (stats.getRowCount() > 0)) {
+              outPathsCommitted[i] = updaters[i].getUpdatedFilePath();
+            }
+            updaters[i].close(abort); 
           }
         }
       } catch (IOException e) {
@@ -249,7 +267,9 @@ public class FileSinkOperator extends TerminalOperator<FileSinkDesc> implements
         if (isMmTable) {
           assert outPaths[idx].equals(finalPaths[idx]);
           commitPaths.add(outPaths[idx]);
-        } else if (!fs.rename(outPaths[idx], finalPaths[idx])) {
+        } else if (isDirectInsert && (outPathsCommitted[idx] != null)) {
+          commitPaths.add(outPathsCommitted[idx]);
+        } else if (!isDirectInsert && !fs.rename(outPaths[idx], finalPaths[idx])) {
             FileStatus fileStatus = FileUtils.getFileStatusOrNull(fs, finalPaths[idx]);
             if (fileStatus != null) {
               LOG.warn("Target path " + finalPaths[idx] + " with a size " + fileStatus.getLen() + " exists. Trying to delete it.");
@@ -264,16 +284,29 @@ public class FileSinkOperator extends TerminalOperator<FileSinkDesc> implements
             }
         }
       }
-
       updateProgress();
     }
 
-    public void abortWriters(FileSystem fs, boolean abort, boolean delete) throws HiveException {
-      //should this close updaters[]?
+    public void abortWritersAndUpdaters(FileSystem fs, boolean abort, boolean delete) throws HiveException {
       for (int idx = 0; idx < outWriters.length; idx++) {
         if (outWriters[idx] != null) {
           try {
+            LOG.debug("Aborted: closing: " + outWriters[idx].toString());
             outWriters[idx].close(abort);
+            if (delete) {
+              fs.delete(outPaths[idx], true);
+            }
+            updateProgress();
+          } catch (IOException e) {
+            throw new HiveException(e);
+          }
+        }
+      }
+      for (int idx = 0; idx < updaters.length; idx++) {
+        if (updaters[idx] != null) {
+          try {
+            LOG.debug("Aborted: closing: " + updaters[idx].toString());
+            updaters[idx].close(abort);
             if (delete) {
               fs.delete(outPaths[idx], true);
             }
@@ -290,7 +323,7 @@ public class FileSinkOperator extends TerminalOperator<FileSinkDesc> implements
       if (isNativeTable) {
         String extension = Utilities.getFileExtension(jc, isCompressed, hiveOutputFormat);
         String taskWithExt = extension == null ? taskId : taskId + extension;
-        if (!isMmTable) {
+        if (!isMmTable && !isDirectInsert) {
           if (!bDynParts && !isSkewedStoredAsSubDirectories) {
             finalPaths[filesIdx] = new Path(parent, taskWithExt);
           } else {
@@ -307,8 +340,13 @@ public class FileSinkOperator extends TerminalOperator<FileSinkDesc> implements
           if (extension != null) {
             taskIdPath += extension;
           }
-
-          Path finalPath = new Path(buildTmpPath(), taskIdPath);
+          
+          Path finalPath;
+          if (isDirectInsert) {
+            finalPath = buildTmpPath();
+          } else {
+            finalPath = new Path(buildTmpPath(), taskIdPath);
+          }
 
           // In the cases that have multi-stage insert, e.g. a "hive.skewjoin.key"-based skew join,
           // it can happen that we want multiple commits into the same directory from different
@@ -319,7 +357,7 @@ public class FileSinkOperator extends TerminalOperator<FileSinkDesc> implements
           // affects some less obscure scenario.
           try {
             FileSystem fpfs = finalPath.getFileSystem(hconf);
-            if (fpfs.exists(finalPath)) {
+            if ((!isDirectInsert) && fpfs.exists(finalPath)) {
               throw new RuntimeException(finalPath + " already exists");
             }
           } catch (IOException e) {
@@ -330,7 +368,7 @@ public class FileSinkOperator extends TerminalOperator<FileSinkDesc> implements
         }
         if (LOG.isInfoEnabled()) {
           LOG.info("Final Path: FS " + finalPaths[filesIdx]);
-          if (LOG.isInfoEnabled() && !isMmTable) {
+          if (LOG.isInfoEnabled() && (!isMmTable && !isDirectInsert)) {
             LOG.info("Writing to temp file: FS " + outPaths[filesIdx]);
           }
         }
@@ -468,10 +506,13 @@ public class FileSinkOperator extends TerminalOperator<FileSinkDesc> implements
       unionPath = null;
     } else {
       isUnionDp = (dpCtx != null);
-      if (conf.isMmTable() || isUnionDp) {
+      if (conf.isMmTable() || isUnionDp) { 
         // MM tables need custom handling for union suffix; DP tables use parent too.
         specPath = conf.getParentDir();
         unionPath = conf.getDirName().getName();
+      } else if (conf.isDirectInsert()) {
+        specPath = conf.getParentDir();
+        unionPath = null;
       } else {
         // For now, keep the old logic for non-MM non-DP union case. Should probably be unified.
         specPath = conf.getDirName();
@@ -526,7 +567,11 @@ public class FileSinkOperator extends TerminalOperator<FileSinkDesc> implements
         throw ex;
       }
       isCompressed = conf.getCompressed();
-      parent = Utilities.toTempPath(conf.getDirName());
+      if (conf.isLinkedFileSink() && conf.isDirectInsert()) {
+        parent = Utilities.toTempPath(conf.getFinalDirName());
+      } else {
+        parent = Utilities.toTempPath(conf.getDirName());
+      }
       statsFromRecordWriter = new boolean[numFiles];
       serializer = (Serializer) conf.getTableInfo().getDeserializerClass().newInstance();
       serializer.initialize(unsetNestedColumnPaths(hconf), conf.getTableInfo().getProperties());
@@ -565,7 +610,7 @@ public class FileSinkOperator extends TerminalOperator<FileSinkDesc> implements
       }
 
       if (!bDynParts) {
-        fsp = new FSPaths(specPath, conf.isMmTable());
+        fsp = new FSPaths(specPath, conf.isMmTable(), conf.isDirectInsert());
         fsp.subdirAfterTxn = combinePathFragments(generateListBucketingDirName(null), unionPath);
         if (Utilities.FILE_OP_LOGGER.isTraceEnabled()) {
           Utilities.FILE_OP_LOGGER.trace("creating new paths " + System.identityHashCode(fsp)
@@ -740,7 +785,7 @@ public class FileSinkOperator extends TerminalOperator<FileSinkDesc> implements
       assert filesIdx == numFiles;
 
       // in recent hadoop versions, use deleteOnExit to clean tmp files.
-      if (isNativeTable() && fs != null && fsp != null && !conf.isMmTable()) {
+      if (isNativeTable() && fs != null && fsp != null && !conf.isMmTable() && !conf.isDirectInsert()) {
         autoDelete = fs.deleteOnExit(fsp.outPaths[0]);
       }
     } catch (Exception e) {
@@ -764,7 +809,7 @@ public class FileSinkOperator extends TerminalOperator<FileSinkDesc> implements
         LOG.info("New Final Path: FS " + fsp.finalPaths[filesIdx]);
       }
 
-      if (isNativeTable() && !conf.isMmTable()) {
+      if (isNativeTable() && !conf.isMmTable() && !conf.isDirectInsert()) {
         // in recent hadoop versions, use deleteOnExit to clean tmp files.
         autoDelete = fs.deleteOnExit(fsp.outPaths[filesIdx]);
       }
@@ -779,7 +824,7 @@ public class FileSinkOperator extends TerminalOperator<FileSinkDesc> implements
       // If MM wants to create a new base for IOW (instead of delta dir), it should specify it here
       if (conf.getWriteType() == AcidUtils.Operation.NOT_ACID || conf.isMmTable()) {
         Path outPath = fsp.outPaths[filesIdx];
-        if (conf.isMmTable()
+        if (conf.isMmTable() 
             && !FileUtils.mkdir(fs, outPath.getParent(), hconf)) {
           LOG.warn("Unable to create directory with inheritPerms: " + outPath);
         }
@@ -790,12 +835,18 @@ public class FileSinkOperator extends TerminalOperator<FileSinkDesc> implements
             StatsProvidingRecordWriter;
         // increment the CREATED_FILES counter
       } else if (conf.getWriteType() == AcidUtils.Operation.INSERT) {
+        Path outPath = fsp.outPaths[filesIdx];
+        if (conf.isDirectInsert() 
+            && !FileUtils.mkdir(fs, outPath.getParent(), hconf)) {
+          LOG.warn("Unable to create directory with inheritPerms: " + outPath);
+        }
         // Only set up the updater for insert.  For update and delete we don't know unitl we see
         // the row.
         ObjectInspector inspector = bDynParts ? subSetOI : outputObjInspector;
         int acidBucketNum = Integer.parseInt(Utilities.getTaskIdFromFilename(taskId));
         fsp.updaters[filesIdx] = HiveFileFormatUtils.getAcidRecordUpdater(jc, conf.getTableInfo(),
-            acidBucketNum, conf, fsp.outPaths[filesIdx], inspector, reporter, -1);
+            acidBucketNum, conf, fsp.outPaths[filesIdx], inspector, reporter, -1); // outPath.getParent()
+        
       }
 
       if (reporter != null) {
@@ -824,9 +875,9 @@ public class FileSinkOperator extends TerminalOperator<FileSinkDesc> implements
       // <table-dir>/<staging-dir>/<partition-dir>/<union-dir>
 
       // for non-MM tables, the final destination partition directory is created during move task via rename
-      // for MM tables, the final destination partition directory is created by the tasks themselves
+      // for MM tables and ACID insert, the final destination partition directory is created by the tasks themselves
       try {
-        if (conf.isMmTable()) {
+        if (conf.isMmTable() || conf.isDirectInsert()) {
           createDpDir(destPartPath);
         } else {
           // outPath will be
@@ -1086,7 +1137,7 @@ public class FileSinkOperator extends TerminalOperator<FileSinkDesc> implements
    * @throws HiveException
    */
   private FSPaths createNewPaths(String dpDir, String lbDir) throws HiveException {
-    FSPaths fsp2 = new FSPaths(specPath, conf.isMmTable());
+    FSPaths fsp2 = new FSPaths(specPath, conf.isMmTable(), conf.isDirectInsert());
     fsp2.subdirAfterTxn = combinePathFragments(lbDir, unionPath);
     fsp2.subdirBeforeTxn = dpDir;
     String pathKey = combinePathFragments(dpDir, lbDir);
@@ -1306,7 +1357,7 @@ public class FileSinkOperator extends TerminalOperator<FileSinkDesc> implements
         // record writer already gathers the statistics, it can simply return the
         // accumulated statistics which will be aggregated in case of spray writers
         if (conf.isGatherStats() && isCollectRWStats) {
-          if (conf.getWriteType() == AcidUtils.Operation.NOT_ACID || conf.isMmTable()) {
+          if (conf.getWriteType() == AcidUtils.Operation.NOT_ACID || conf.isMmTable()) { 
             for (int idx = 0; idx < fsp.outWriters.length; idx++) {
               RecordWriter outWriter = fsp.outWriters[idx];
               if (outWriter != null) {
@@ -1337,9 +1388,9 @@ public class FileSinkOperator extends TerminalOperator<FileSinkDesc> implements
           SparkMetricUtils.updateSparkBytesWrittenMetrics(LOG, fs, fsp.finalPaths);
         }
       }
-      if (conf.isMmTable()) {
-        Utilities.writeMmCommitManifest(commitPaths, specPath, fs, originalTaskId,
-                conf.getTableWriteId(), conf.getStatementId(), unionPath, conf.getInsertOverwrite());
+      if (conf.isMmTable() || conf.isDirectInsert()) {
+        Utilities.writeCommitManifest(commitPaths, specPath, fs, originalTaskId, conf.getTableWriteId(), conf
+            .getStatementId(), unionPath, conf.getInsertOverwrite());
       }
       // Only publish stats if this operator's flag was set to gather stats
       if (conf.isGatherStats()) {
@@ -1350,7 +1401,7 @@ public class FileSinkOperator extends TerminalOperator<FileSinkDesc> implements
       // Hadoop always call close() even if an Exception was thrown in map() or
       // reduce().
       for (FSPaths fsp : valToPaths.values()) {
-        fsp.abortWriters(fs, abort, !autoDelete && isNativeTable() && !conf.isMmTable());
+        fsp.abortWritersAndUpdaters(fs, abort, !autoDelete && isNativeTable() && !conf.isMmTable());
       }
     }
     fsp = prevFsp = null;
@@ -1383,10 +1434,14 @@ public class FileSinkOperator extends TerminalOperator<FileSinkDesc> implements
           specPath = conf.getParentDir();
           unionSuffix = conf.getDirName().getName();
         }
+        if (conf.isLinkedFileSink() && conf.isDirectInsert()) {
+          specPath = conf.getParentDir();
+          unionSuffix = null;
+        }
         if (Utilities.FILE_OP_LOGGER.isTraceEnabled()) {
           Utilities.FILE_OP_LOGGER.trace("jobCloseOp using specPath " + specPath);
         }
-        if (!conf.isMmTable()) {
+        if (!conf.isMmTable() && !conf.isDirectInsert()) {
           Utilities.mvFileToFinalPath(specPath, hconf, success, LOG, dpCtx, conf, reporter);
         } else {
           int dpLevels = dpCtx == null ? 0 : dpCtx.getNumDPCols(),
@@ -1396,9 +1451,9 @@ public class FileSinkOperator extends TerminalOperator<FileSinkDesc> implements
               : (dpCtx != null ? dpCtx.getNumBuckets() : 0);
           MissingBucketsContext mbc = new MissingBucketsContext(
               conf.getTableInfo(), numBuckets, conf.getCompressed());
-          Utilities.handleMmTableFinalPath(specPath, unionSuffix, hconf, success,
-              dpLevels, lbLevels, mbc, conf.getTableWriteId(), conf.getStatementId(), reporter,
-              conf.isMmTable(), conf.isMmCtas(), conf.getInsertOverwrite());
+          Utilities.handleDirectInsertTableFinalPath(specPath, unionSuffix, hconf, success, dpLevels, lbLevels, mbc,
+              conf.getTableWriteId(), conf.getStatementId(), reporter, conf.isMmTable(), conf.isMmCtas(), conf
+                  .getInsertOverwrite(), conf.isDirectInsert());
         }
       }
     } catch (IOException e) {
